requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
rank_bm25==0.2.2
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
transformers==4.44.2
numpy==1.26.4
scikit-learn==1.5.1
tqdm==4.66.4
python-dotenv==1.0.1
